extract searcher main to shared package for external invocation
This is a copy-and-paste of the searcher main function to a shared package so that searcher can be invoked from a different entrypoint. This is how it's done for several other programs (frontend, repo-updater, gitserver, etc.). There are no logic changes.
